webui: slightly reduce wsRegister() overhead
Since there are no dynamic callbacks registered at setup(),
use simple function pointers as the input type.

Internal type is still std::func, though
(and not yet sure how to get rid of it correctly)

+++ code/espurna/debug.cpp
@@ -186,7 +186,7 @@ void formatAndSend(const char* format, va_list args) {
         return;
     }
 
-    const size_t BufferSize { len + 1 };
+    const size_t BufferSize { static_cast<size_t>(len) + 1 };
     auto* buffer = new (std::nothrow) char[BufferSize];
     if (!buffer) {
         return;
